+++ Multiplayer/Scripts/Console.cs
@@ -1,0 +1,8 @@
+﻿using UnityEngine;
+public static class Console
+{
+    public static void Log(object message)
+    {
+        Debug.Log("Multiplayer Mod: " + message);
+    }
+}
