+++ Multiplayer/Scripts/Tags.cs
@@ -1,0 +1,11 @@
+﻿public enum Tags
+{
+    SpawnPlayerTag,
+    PlayerHandLeft_Movement, PlayerHandLeft_Rotation,
+    PlayerHandRight_Movement, PlayerHandRight_Rotation,
+    PlayerHead_Movement, PlayerHead_Rotation,
+    ServerInfo,
+    DestroyPlayer,
+    PlayersInfo,
+    BasicVehicle_Movement, BasicVehicle_Rotation
+}
